requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pdfkit==1.0.0
wkhtmltopdf==0.2
requests==2.31.0